Refactor header into own component
